seperate text-list and indicator container
